# swagger-peminjaman-ruang/swagger.yaml
openapi: 3.1.3
info:
  title: SIANGKASA-API
  description: >-
    Setiap request yang ada, harus selalu menyertakan 2 header


    1. gateway-key = berisikan kunci gateway yang ada pada environment project
    api-gateway-key (untuk development menggunakan DEV_GATEWAY, untuk production
    menggunakan API_GATEWAY)

    2. unixtime = berisikan konversi waktu request ke dalam unixtime untuk di
    validasi saat production (saat development, unixtime bisa diisi angka apa
    saja karena tidak akan di validasi)


    Optional:


    Untuk API yang terdapat simbol kunci, baik secara global maupun
    individu, wajib menyertakan header


    1. Authorization = berisikan token yang di dapat saat melakukan login


    Authorization ini dapat berbentuk "Authorization: {{ key }}" atau
    "Authorization: Bearer {{ key }}"
  version: 1.0.0
servers:
  - url: http://localhost:9002
...
components:
  securitySchemes:
    Authorization:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /email/sent:
    post:
      tags:
        - Email Service
      summary: Sent Email
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              properties:
                emailTo:
                  type: string
                  example: test@example.com
                subject:
                  type: string
                  example: TEST
                body:
                  type: string
                  example: Test Krim Email
      responses:
        "200":
          description: Succesful Response
          content:
            application/json: {}
        "400":
          description: Request Tidak Lengkap
          content:
            application/json: {}
        "403":
          description: Forbidden
          content:
            application/json: {}
        "404":
          description: Data Tidak Ditemukan
          content:
            application/json: {}
        "500":
          description: Error
          content:
            application/json: {}
...